Rename manage_oidc to manage_keycloak
