+++ flask/app/__init__.py
@@ -16,7 +16,6 @@ from app import (
     groups,
     users,
     manage,
-    manage_oidc,
     error_handler,
 )
